requirements.txt
flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
matplotlib==3.7.2
deepface==0.0.79
tensorflow==2.13.0
google-api-python-client==2.93.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.22.0
gunicorn==21.2.0